refactor(user-registration): changes on PR comments
